🔨 [#10] 자산 관련 타입 Integer -> Long으로 수정

+++ src/main/java/com/example/demo/room/dto/res/RoomRes.java
@@ -13,7 +13,7 @@ public class RoomRes {
     public record CreateRoom(
             Long roomId, Integer maxMember, String pwd, String inviteCode,
             Integer maxTeam, Integer maxRound, RoomStatus roomStatus, Mode mode,
-            String yearSet, Integer seedMoney, String nickName) {}
+            String yearSet, Long seedMoney, String nickName) {}
 
     public record JoinRoom(Long roomId, String nickName, List<TeamRes.TeamDetail> teamDetail) {}
